exporting folder only if folder contains data

+++ goobi1.9/WEB-INF/src/de/sub/goobi/Export/dms/AutomaticDmsExport.java
@@ -102,8 +102,7 @@ public class AutomaticDmsExport extends ExportMets {
 		String atsPpnBand = myProzess.getTitel();
 
 		/*
-		 * -------------------------------- Dokument einlesen
-		 * --------------------------------
+		 * -------------------------------- Dokument einlesen --------------------------------
 		 */
 		Fileformat gdzfile;
 		Fileformat newfile;
@@ -135,8 +134,7 @@ public class AutomaticDmsExport extends ExportMets {
 		trimAllMetadata(gdzfile.getDigitalDocument().getLogicalDocStruct());
 
 		/*
-		 * -------------------------------- Metadaten validieren
-		 * --------------------------------
+		 * -------------------------------- Metadaten validieren --------------------------------
 		 */
 
 		if (ConfigMain.getBooleanParameter("useMetadatenvalidierung")) {
@@ -147,8 +145,7 @@ public class AutomaticDmsExport extends ExportMets {
 		}
 
 		/*
-		 * -------------------------------- Speicherort vorbereiten und
-		 * downloaden --------------------------------
+		 * -------------------------------- Speicherort vorbereiten und downloaden --------------------------------
 		 */
 		String zielVerzeichnis;
 		File benutzerHome;
@@ -184,8 +181,7 @@ public class AutomaticDmsExport extends ExportMets {
 		}
 
 		/*
-		 * -------------------------------- der eigentliche Download der Images
-		 * --------------------------------
+		 * -------------------------------- der eigentliche Download der Images --------------------------------
 		 */
 		try {
 			if (this.exportWithImages) {
@@ -200,10 +196,8 @@ public class AutomaticDmsExport extends ExportMets {
 		}
 
 		/*
-		 * -------------------------------- zum Schluss Datei an gewünschten Ort
-		 * exportieren entweder direkt in den Import-Ordner oder ins
-		 * Benutzerhome anschliessend den Import-Thread starten
-		 * --------------------------------
+		 * -------------------------------- zum Schluss Datei an gewünschten Ort exportieren entweder direkt in den Import-Ordner oder ins
+		 * Benutzerhome anschliessend den Import-Thread starten --------------------------------
 		 */
 		if (myProzess.getProjekt().isUseDmsImport()) {
 			if (MetadataFormat.findFileFormatsHelperByName(myProzess.getProjekt().getFileFormatDmsExport()) == MetadataFormat.METS) {
@@ -270,8 +264,7 @@ public class AutomaticDmsExport extends ExportMets {
 	}
 
 	/**
-	 * run through all metadata and children of given docstruct to trim the
-	 * strings calls itself recursively
+	 * run through all metadata and children of given docstruct to trim the strings calls itself recursively
 	 */
 	private void trimAllMetadata(DocStruct inStruct) {
 		/* trimm all metadata values */
@@ -299,25 +292,23 @@ public class AutomaticDmsExport extends ExportMets {
 
 		// download sources
 		File sources = new File(myProzess.getSourceDirectory());
-		if (sources.exists()) {
-			File destination = new File(benutzerHome + File.separator
-					+ atsPpnBand + "_src");
+		if (sources.exists() && sources.list().length > 0) {
+			File destination = new File(benutzerHome + File.separator + atsPpnBand + "_src");
 			if (!destination.exists()) {
 				destination.mkdir();
 			}
 			File[] dateien = sources.listFiles();
 			for (int i = 0; i < dateien.length; i++) {
-				File meinZiel = new File(destination + File.separator
-						+ dateien[i].getName());
+				File meinZiel = new File(destination + File.separator + dateien[i].getName());
 				Helper.copyFile(dateien[i], meinZiel);
 			}
 		}
-		
+
 		File ocr = new File(myProzess.getOcrDirectory());
 		if (ocr.exists()) {
 			File[] folder = ocr.listFiles();
 			for (File dir : folder) {
-				if (dir.isDirectory()) {
+				if (dir.isDirectory() && dir.list().length > 0) {
 					String suffix = dir.getName().substring(dir.getName().lastIndexOf("_"));
 					File destination = new File(benutzerHome + File.separator + atsPpnBand + "_" + suffix);
 					if (!destination.exists()) {
@@ -331,54 +322,51 @@ public class AutomaticDmsExport extends ExportMets {
 				}
 			}
 		}
-		
-		
-		
-//		File txtFolder = new File(myProzess.getTxtDirectory());
-//		if (txtFolder.exists()) {
-//			File destination = new File(benutzerHome + File.separator + atsPpnBand + "_txt");
-//			if (!destination.exists()) {
-//				destination.mkdir();
-//			}
-//			File[] dateien = txtFolder.listFiles();
-//			for (int i = 0; i < dateien.length; i++) {
-//				File meinZiel = new File(destination + File.separator + dateien[i].getName());
-//				Helper.copyFile(dateien[i], meinZiel);
-//			}
-//		}
-//
-//		File wordFolder = new File(myProzess.getWordDirectory());
-//		if (wordFolder.exists()) {
-//			File destination = new File(benutzerHome + File.separator + atsPpnBand + "_wc");
-//			if (!destination.exists()) {
-//				destination.mkdir();
-//			}
-//			File[] dateien = wordFolder.listFiles();
-//			for (int i = 0; i < dateien.length; i++) {
-//				File meinZiel = new File(destination + File.separator + dateien[i].getName());
-//				Helper.copyFile(dateien[i], meinZiel);
-//			}
-//		}
-//
-//		File pdfFolder = new File(myProzess.getPdfDirectory());
-//		if (pdfFolder.exists()) {
-//			File destination = new File(benutzerHome + File.separator + atsPpnBand + "_pdf");
-//			if (!destination.exists()) {
-//				destination.mkdir();
-//			}
-//			File[] dateien = pdfFolder.listFiles();
-//			for (int i = 0; i < dateien.length; i++) {
-//				File meinZiel = new File(destination + File.separator + dateien[i].getName());
-//				Helper.copyFile(dateien[i], meinZiel);
-//			}
-//		}
+
+		// File txtFolder = new File(myProzess.getTxtDirectory());
+		// if (txtFolder.exists()) {
+		// File destination = new File(benutzerHome + File.separator + atsPpnBand + "_txt");
+		// if (!destination.exists()) {
+		// destination.mkdir();
+		// }
+		// File[] dateien = txtFolder.listFiles();
+		// for (int i = 0; i < dateien.length; i++) {
+		// File meinZiel = new File(destination + File.separator + dateien[i].getName());
+		// Helper.copyFile(dateien[i], meinZiel);
+		// }
+		// }
+		//
+		// File wordFolder = new File(myProzess.getWordDirectory());
+		// if (wordFolder.exists()) {
+		// File destination = new File(benutzerHome + File.separator + atsPpnBand + "_wc");
+		// if (!destination.exists()) {
+		// destination.mkdir();
+		// }
+		// File[] dateien = wordFolder.listFiles();
+		// for (int i = 0; i < dateien.length; i++) {
+		// File meinZiel = new File(destination + File.separator + dateien[i].getName());
+		// Helper.copyFile(dateien[i], meinZiel);
+		// }
+		// }
+		//
+		// File pdfFolder = new File(myProzess.getPdfDirectory());
+		// if (pdfFolder.exists()) {
+		// File destination = new File(benutzerHome + File.separator + atsPpnBand + "_pdf");
+		// if (!destination.exists()) {
+		// destination.mkdir();
+		// }
+		// File[] dateien = pdfFolder.listFiles();
+		// for (int i = 0; i < dateien.length; i++) {
+		// File meinZiel = new File(destination + File.separator + dateien[i].getName());
+		// Helper.copyFile(dateien[i], meinZiel);
+		// }
+		// }
 	}
 
 	public void imageDownload(Prozess myProzess, File benutzerHome, String atsPpnBand, final String ordnerEndung) throws IOException,
 			InterruptedException, SwapException, DAOException {
 		/*
-		 * -------------------------------- erstmal alle Filter
-		 * --------------------------------
+		 * -------------------------------- erstmal alle Filter --------------------------------
 		 */
 		// FilenameFilter filterTifDateien = new FilenameFilter() {
 		// public boolean accept(File dir, String name) {
@@ -387,17 +375,15 @@ public class AutomaticDmsExport extends ExportMets {
 		// };
 
 		/*
-		 * -------------------------------- dann den Ausgangspfad ermitteln
-		 * --------------------------------
+		 * -------------------------------- dann den Ausgangspfad ermitteln --------------------------------
 		 */
 		Helper help = new Helper();
 		File tifOrdner = new File(myProzess.getImagesTifDirectory());
 
 		/*
-		 * -------------------------------- jetzt die Ausgangsordner in die
-		 * Zielordner kopieren --------------------------------
+		 * -------------------------------- jetzt die Ausgangsordner in die Zielordner kopieren --------------------------------
 		 */
-		if (tifOrdner.exists()) {
+		if (tifOrdner.exists() && tifOrdner.list().length > 0) {
 			File zielTif = new File(benutzerHome + File.separator + atsPpnBand + ordnerEndung);
 
 			/* bei Agora-Import einfach den Ordner anlegen */
@@ -407,8 +393,7 @@ public class AutomaticDmsExport extends ExportMets {
 				}
 			} else {
 				/*
-				 * wenn kein Agora-Import, dann den Ordner mit
-				 * Benutzerberechtigung neu anlegen
+				 * wenn kein Agora-Import, dann den Ordner mit Benutzerberechtigung neu anlegen
 				 */
 				Benutzer myBenutzer = (Benutzer) Helper.getManagedBeanValue("#{LoginForm.myBenutzer}");
 				try {
